Updated openstack/openstack Project: openstack/horizon  3c32807843f9ec56c69b27523ad75d905bb3b2ee
Fix popup error when volume service disabled

In an enviroment, where cinder is disabled, I'm getting error popups
when trying to 'launch instance':
Error: Unable to retrieve list of volumes.
Error: Unable to retrieve list of volume snapshots.
So we need to make sure whether the volume service exists when
calling the cinder api in the launch instance workflow.

Closes-bug: #1394900

Change-Id: I6f745f7f95272eb87b6d867d76d7726c7eb5bb24