Updated openstack/openstack Project: openstack-infra/infra-manual  b2ad3911135006ea526706f9cd54bdb63d461bf7
Adds a link for new developers to use sandbox

Right now the docs for new developers don't include instructions
for getting started with openstack-dev/sandbox. This patches adds
a link to a tutorial for working with the sandbox.

It also generalizes the instructions for the Starting Work on
a New Project section, using <projectname> instead of nova.

Change-Id: Ied45a74f8573b04e662f45168c29b56f7eaab33c