Updated openstack/openstack Project: openstack/taskflow  c714e29c28683373ed5d333dba064ace4ebaf139
Add a claims listener that connects job claims to engines

To make it easily possible to stop running a engine that was
created from a job, add a claims listener that will be called
on state changes that an engine progresses through. During those
state changes the jobboard will be queried to determine if the
job is still claimed by the respective owner; if not the engine
will be suspended and further work will stop.

Change-Id: I8bbc6a3e03746ba0a7c74139cf9e230631d80d8f