Updated openstack/openstack Project: openstack/neutron  421650230a006dc3bd1269aa8772ce5d1fb7bb95
iptables_manager: Fix get_binary_name for eventlet

The inspect trick used by get_binary_name does not work for
eventlet non-main threads.  It's actually the case for ofagent
and in that case it yields something like "greenthread.py",
which is not desirable because this module assumes it different
for each programs.  This commit fixes the problem by switching
to use sys.argv.

Closes-Bug: #1367075
Change-Id: I8e765882c58d9991a37f0bbb52cdccc406e287bd