Updated openstack/openstack Project: openstack-dev/devstack  20132c1aeb1c96bae1446e70f9e2ff664b947014
Rename apts to debs

apts aren't a thing, debs are. apt-get installs debs, like yum installs
rpms. Another option would be to rename rpms to yums, but that is even
sillier, considering then for suse we'd have to call it "zypper-suse" or
something, even though it would want to consume yums-general as a base.

Include a symlink to the old name to help out with grenade. The symlink
should go away later.

This is a long-standing annoyance of mine.

Related-to: I915f0e59c13d8cf5fb3549c6202f8aa2f8be0856
Change-Id: I0416180db5b6add996ce5b48c6966c1b68adbcb0