Updated openstack/openstack Project: openstack/nova  e32ae64dda5f5263e6bac7d1c985c52d90a19d30
VMware: support spawn of stream-optimized image

This adds support of spawning instances using stream-optimized
disk images. This format is better suited as a glance format as it is
compressed, sparsed and optimized for streaming. The change uses the
oslo.vmware library to provide the functionality to transfer the image
to the server by importing it as a virtual machine.

Implements: blueprint vmware-vsan-support
Co-authored-by: Radoslav Gerganov <rgerganov@vmware.com>

Change-Id: I5c5c2aae733a2d55643d2ea5c293873b21325db6