Updated openstack/openstack Project: openstack/taskflow  87abb27f1759fdce06c72fcbd2ece69ab06c5ea4
Add a sample script that can be used to build a test environment

It can be difficult to figure out what to install to get a test
environment that can be used to test all the opportunistic tests,
zookeeper and mysql (with the right user and password and test
database).

This script provides a way to do this on rhel/ubuntu based
environments and can be used to examine how to do this for other
environment types.

I use it with nova boot (on rhel images) and it has worked out
great as a way to rebuild a clean test environment as I need.

Change-Id: I9aef7e5536aa38e951a6411a8e83097339a43955