Updated openstack/openstack Project: openstack/nova  342e5d950893162ea5dc435f85754066f877cb47
Drop support for legacy server groups

A previous commit dropped the scheduler filters needed for legacy
server groups to work.  This patch completes removing that support.
If someone attempts to create a server with a group hint that contains
a name instead of a UUID, it will return a proper HTTP failure.

Change-Id: Ic8a92c8ec1f5bbb240803448014757f89facbb4d