Updated openstack/openstack Project: openstack/nova  7bf9b7ed3493628772a3165dc04c3a22c8a1cb34
tests: update _get_instance_xml to accept custom flavor object

Currently LibvirtVifTestCase._get_instance_xml creates a default
flavor object and uses it for all tests. This makes it dificult
to test settings that are being activated by values in extra
spects.

Allowing the method to accept a flavor object and use the default
one, if not provided.

Change-Id: I3835aa5e6cf09268b132b9de377004c0e8df9cb0