Updated openstack/openstack Project: openstack/nova  c3f3dc012ae3938b6f116491273a4eef0acfab83
Break V2 XML Support

Nova's XML support has been deprecated with a warning that it could be
removed as early as Juno. Now that we are in Kilo we are ready to remove
support.  Before actually doing the surgery to properly remove XML
support, intentionally break XML support in a 'clean' way, so if someone
freaks out that we are removing XML support we can easily re-enable it.

Follow up to I4acc15ce5f487738bb34a95c2261a5d05d827d8d

Co-Authored-By: Joe Gordon <joe.gordon0@gmail.com>
Change-Id: I5a580fc323c3809790b4a68a9f8f8129ecdc2cf0