Updated openstack/openstack Project: openstack/nova  e26973d2380e5a8b9777104679ed6db89f4d3c74
Make aggregate filters use objects

Some scheduler filters that use aggregates have been converted
to use nova objects. This patch converts the remaining ones to
use objects. In both cases the calls to the database are
replaced by calls to a utility method that gets the required
data using Aggregate objects. In the future these utility
methods could be replaced with an implementation that works on
a cache - for now they do not change the behavior - its just
a refactor.

blueprint kilo-objects
Change-Id: I42d5697a67805b7e85f82cf3cf2cd6e17d923955