Updated openstack/openstack Project: openstack/nova  c44f919f8d2ba5990a6ade986e0d07c4acc79cba
libvirt: Fix domain creation for LXC

A condition was introduced during domain creation that needs
disk_info to determine whether or not an instance was booted from
volume. Many calls to _create_domain_and_network were not providing
disk_info and thus the condition was always evaluating to True.
This change adds disk_info to each call that was missing it.

Closes-bug: #1370590

Co-Authored-By: Andrew Melton <andrew.melton@rackspace.com>

Change-Id: I23e7e013d3a51b8040dd37e5facfe75b480e1721