Updated openstack/openstack Project: openstack/keystone  7541fda1db8290f639e19420b7ac3f683aab27a5
Add an identity backend method to get group by name.

We already support a get_user_by_name method, but an equivalent
one for groups is missing. This is needed for enhancements to
our federation mapping engine. This patch provides this new
method.

This patch does not expose this functionality as part of the
REST API, it is only supported at the manager/driver level.

Partially implements: bp mapping-enhancements

Change-Id: I2508a99125d9aec2a66b3b48e63c9b1ed2fe12a4