Updated openstack/openstack Project: openstack/nova  805e690372fd13e9e5ddb8dcfd1801682538f8ec
libvirt: Enable console and log for system z guests

If the host architecture is system z (s390, s390x), the interactive
console and the log of instances will be defined as "console" devices
(instead of "serial" devices) in libvirts domain XML. These console
devices need to have a "target type". For the interactive console this
is "sclp" and for the log "sclplm".

Partial-Implements blueprint libvirt-kvm-systemz

Change-Id: I20d5e32fee660e77c52b6fc5fdf1128a996bb994