Updated openstack/openstack Project: openstack/keystonemiddleware  563f82727caee24f5cd8598b0af54d7224d0d4e3
Use real discovery object in auth_token middleware.

Auth token middleware has always done its own version of discovery. It
is slightly more lenient than the one in keystoneclient however we
should use the standard tools for these jobs.

Move the handling of this discovery into the plugin. This is inline with
how auth plugins work and will make it easy to replace with other
plugins.

Change-Id: I3c4ff5bed1aa76dfff8e445de5a384c3f7992bbe