Updated openstack/openstack Project: openstack/taskflow  cb9bc57aa9e36f0c59d94fe1d44b4ab65a34e9ca
Some package additions and adjustments to the env_builder.sh

This just adjusts to install wget and certain other packages
before further work commences to ensure they are ready and
avaiable for further use.

Tested on rhel6.x as a cloud-init userdata script and it appears
to continue to work as expected.

Change-Id: Ia85715e27bc8342fd61d23e2c8a659dbce20aea9