Updated openstack/openstack Project: openstack/nova  5eea055b8b6afa0fc2fb50b065da754553cb71ea
Enforce unique instance uuid in data model

This adds the database migration that checks for records
where instance_uuid is NULL and fails if found until the
'nova-manage db null_instance_uuid_scan' command is run
and the records are manually deleted.

Unit tests are added for the migration to make sure the
fixed_ips table is left alone and the UniqueConstraint is
created on the instances table.

DocImpact: New 'nova-manage db null_instance_uuid_scan'
           command.

UpgradeImpact: The new database migration script.

Implements blueprint enforce-unique-instance-uuid-in-db

Change-Id: I768c36b140c7e9f437907d4dfe175e1410ca44b0