Updated openstack/openstack Project: openstack/nova  45cc1a79fc1680d26a1602f0cc6a4b28d913de56
Convert v3/v2.1 extension info to present v2 API format

Changes from the v3 API /extensions interface format to the
v2 API format. Note that this does not yet list the same
extensions as the v2 api because in v2.1 some have been split
or renamed. This will be addressed in a future patch

Partially implements blueprint v2-on-v3-api

Change-Id: Icaf945b356fcbbe2126343ad01ee58cecd15100f