Updated openstack/openstack Project: openstack/heat  cb4e0adccc3bd50f5d15151e10a1e0f7e9a76f57
Remove heat.config.sample file

This file keeps breaking our gate, because libraries we depend on
release and the generated content changes.

The check has been disabled, but to avoid the sample config becoming
totally stale, remove it and replace with a README which says
how to generate the current sample config.

Change-Id: I1c43e5cf2a7ab6fa388b91fd420392996c368e72