Updated openstack/openstack Project: openstack/sahara  8b3908e450eee37a5e94bcd4808b81bb84b47c95
Don't provide CONF to the AuthProtocol middleware

The conf value that is passed to the AuthProtocol middleware is a
dictionary that is expected to be passed from paste. They are overrides
of oslo.conf not the conf object itself. If we don't pass anything it
will use the standard CONF object.

Closes-Bug: #1398524
Change-Id: Ia53a90f9b632b54a79906f07610c7a1ae1c70218