Updated openstack/openstack Project: openstack/cinder  a0cef11325f4000f118341cd5aa95a1dd807d68b
Remove the cinder.conf.sample file

We removed the conf_uptodate checks however we didn't
remove the cinder.conf.sample file, but we should.

Now that we're not checking it, it's guaranteed that it will
most certainly be "wrong".  So, this patch removes it, we'll
update the README.conf.sample file if/when we have an auto
published version available.

Change-Id: I0a04f82a843ce795ec7518b58a3677ce3f2a314d