Updated openstack/openstack Project: openstack/keystone  58edd442316b5ab459350a9b717248520980851f
Typo in policy call

The callback notification to endpoint-policy upon deletion of a policy
had a typo in it. This patch fixes this, and adds a test to ensure this
notification is working as designed. The other three notifications
already have tests, for some reason this one had been missed out.

Closes-Bug: 1399857

Change-Id: I5060c30cd03b3154838db9e6f8fee94110751745