Updated openstack/openstack Project: openstack/sahara  3e1fbe3f48ddd76ac5272d7b13395e5f3072d2a5
Specify CDH version

In plugin.py we give "5" and "5.2.0" choices for users to
select. In deploy.py we will use this value as the fullVersion
parameter passed to create_cluster. We treate "5" as "5.0.0".
Now for there is no further operation difference between 5.0.0
and 5.2.0, we do not do other changes, and do not create
directories or use different classes implementations like for
vanilla.

Closes-Bug: #1397155

Change-Id: Ib064eaf60eef0bf338b9757e338ec61b80e6e812