Updated openstack/openstack Project: openstack/nova  dccd1d77f326ea74c16b29523546817fe1bb6b71
Update LVM lockfile name identical to RAW and Qcow

Currently Raw and Qcow2 use filename as lock file identity
while LVM use whole path. This patch change the lockfile
name of LVM to identical to the others.
It will be helpful to allow driver layer code to remove
those files when the life cycle of an instance ends.

Related-Bug: #1256306

Change-Id: I07aa54032026ca479fe8de08308dae51527a6b13