Updated openstack/openstack Project: openstack/tempest  40545e19d8145df21b31ed143522f02619ebe50e
Ironic: Remove few misguided negative tests on ports

As part of https://bugs.launchpad.net/ironic/+bug/1398350 Ironic is fixing
a bug where the josn-fields in the resources only accepted string and
int, as part of the work Ironic will now support any json serializable
value on those fields.

Some negative tests on ports were considering float to be a malformed
value, which they are not. This patch is removing those tests.

This patch is also adding tests checking for the new supported values
in the API.

The patch in Ironic adding support for multiple types is this one:
https://review.openstack.org/#/c/137762/

Change-Id: I02e95d77a90b3c1a8e203cf83cc0ec28b5460e58