Updated openstack/openstack Project: openstack/taskflow  0960b4e0a93adc7d179b992a89082cbb2775d4a9
Stop returning atoms from execute/revert methods

It is not needed to return the atom that was executed from the
futures result() method, since we can just as easily set an
attribute on the future and reference it from there when using
it later. This is also required for a process based executor since
it is not typically possible to send back a raw task object (and
is not desireable to require this); even if it was possible the
task would be pickled and unpickled multiple times so when this
happens it can not be guaranteed to even be the same object (in
fact it is not).

Part of blueprint process-executor

Change-Id: I4a05ea5dcdef97218312e3a88ed4a1dfdf1b1edf