Updated openstack/openstack Project: openstack/horizon  f7c8e1dda6c1a2bff6ea37ba09560263f222bd2e
Improving more projects link in context selection

The existing icon is a cog which implies making a settings change.
Changing to a list icon to match more closely what is achieved by
the link.

Also, adding the icon to the link rather than having it separate,
this is more consistent with the rest of the UI.

Closes-Bug: #1398515
Change-Id: I10c998c939d47b0a8ef6bbd13c66bddcba12b14b