Updated openstack/openstack Project: openstack/horizon  8f2903ed2c17d89774e7e8fd91ef37c2013a706b
wrong message while deleting non empty pseudo folder

while try to delete a pseudo folder which contains some objects is
showing the success message but the folder is not actually deleted

added a check similar to delete container, the new check will
verify the pseudo folder is empty before call the swift client
call. if the folder is not empty it will throw a exception
with a appropriate error message and the message will be shown to
the user in the pop up message box.

Change-Id: I0bbea6ef6ac26ef357240610535b9d84053a89c2
Closes-Bug: #1347598