Updated openstack/openstack Project: openstack-dev/devstack  b23cdf294bb399dfea9f00d0ee9d558e814f93bb
Configure boto.cfg in tempest to point to SSL bundle

The boto client uses the ca_certificate_files value in the
Boto section of /etc/boto.cfg when making an SSL client
connection. Configure that to be the devstack SSL CA bundle
so the boto tests will pass when running against an SSL or
TLS-proxy configured devstack environment.

Change-Id: I55ae6d6e0c5b3adb4370ec93f40ae1cd47741457