Updated openstack/openstack Project: openstack-dev/devstack  f8cf2777d551fc2c87ad72fd21242411998674c6
XenAPI: Add another plugin directory

Buildroot on 64-bit systems may use /usr/lib64 to store the XAPI plugins
Add this as an option to search for.

The list is getting of acceptable paths is getting longer but some work is going
on in XAPI to allow us to query for this path, which will mean we can get
rid of this list in future.

Change-Id: I79aafb6a86032c7ab04937c9e9bec08661ecdefa