Updated openstack/openstack Project: openstack/requirements  cc7770e1354fc2e5c1b283d476f1f32ca2da5775
Bump up oslo.utils to 1.1.0

Oslo.utils was recently released with version 1.1.0

That release contains some code that was extracted
from taskflow to provide general functionality to
other projects. So to allow taskflow to remove that
code from its own project we need to bump this up
to a version that includes that code.

Change-Id: Icfddc58933288576e7832bfc2bb1902c4f8b5cbc