Updated openstack/openstack Project: openstack/heat  5e940450e0db261775ee3f5a31e6711e2dacc772
Use kwargs for ResourcePropertyConflict exception

Current approach works correct when error raised in heat-engine,
but when oslo-messaging try to send this error to heat-api it calls
this Exception again without any arguments. So it leads to empty output
in message.
This patch adds kwargs properties for directly sending parameters to
parent class and storing them for substitution during re-call in
oslo-messanging.

Change-Id: I4e823206c3a88d27241532d9f393c07208af8b73
Closes-Bug: #1397002